@@ -2,7 +2,7 @@ package ikemura.com.simanchu_takara_sagashi_android.ui.list
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
